requirements.txt
# ===========================
# 基础库与工具
# ===========================
colorama==0.4.6         # 终端彩色输出
tqdm==4.67.3            # 进度条显示
setuptools==82.0.0      # Python包管理工具
wheel==0.46.3           # 打包工具
packaging==24.2         # 版本和依赖解析
attrs==25.4.0           # 属性类工具
six==1.17.0             # Python2/3兼容工具
distro==1.9.0           # Linux发行版识别

# ===========================
# 网络请求与异步相关
# ===========================
requests==2.32.5        # HTTP请求库
requests-toolbelt==1.0.0 # requests工具扩展
aiohttp==3.13.3         # 异步HTTP客户端
aiohappyeyeballs==2.6.1 # 异步DNS解析
aiosignal==1.4.0        # 异步信号处理
httpx==0.28.1           # 支持同步/异步的HTTP客户端
httpcore==1.0.9         # httpx底层库
httpx-sse==0.4.3        # httpx的SSE支持
websocket-client==1.9.0 # WebSocket客户端
multidict==6.7.1        # 多值字典，aiohttp依赖
yarl==1.23.0            # URL处理，aiohttp依赖
frozenlist==1.8.0       # 冻结列表，aiohttp依赖
sniffio==1.3.1          # 异步环境检测
anyio==4.12.1           # 异步并发库

# ===========================
# Web框架与API相关
# ===========================
fastapi==0.135.1        # Web API框架
starlette==0.52.1       # FastAPI底层ASGI框架
uvicorn==0.41.0         # ASGI服务器
h11==0.16.0             # HTTP协议实现，uvicorn依赖
itsdangerous==2.2.0     # 安全cookie/签名
PyYAML==6.0.3           # YAML配置解析

# ===========================
# 数据库与ORM
# ===========================
SQLAlchemy==2.0.48      # ORM数据库操作
mysql-connector-python==9.4.0 # MySQL连接器

# ===========================
# 数据处理与科学计算
# ===========================
numpy==2.4.2            # 数值计算
python-dateutil==2.9.0.post0 # 日期处理
orjson==3.11.7          # 高性能JSON解析
dataclasses-json==0.6.7 # dataclass与JSON互转
marshmallow==3.26.2     # 数据序列化与校验
regex==2026.2.28        # 正则表达式增强

# ===========================
# 类型与序列化相关
# ===========================
pydantic==2.11.10       # 数据校验与类型管理
pydantic-settings==2.13.1 # 配置管理
pydantic_core==2.33.2   # pydantic核心库
typing_extensions==4.15.0 # 类型扩展
typing-inspect==0.9.0   # 类型检查
typing-inspection==0.4.2 # 类型检查
mypy_extensions==1.1.0  # mypy类型扩展

# ===========================
# 加密与安全
# ===========================
cryptography==46.0.5    # 加密库
certifi==2026.2.25      # CA证书
cffi==2.0.0             # C扩展接口
pycparser==3.0          # C代码解析

# ===========================
# LLM/AI/嵌入/知识库相关
# ===========================
openai==2.24.0          # OpenAI API
dashscope==1.25.13      # 阿里云大模型API
langchain==1.2.10       # LLM链式调用
langchain-core==1.2.17  # langchain核心
langchain-classic==1.0.1 # langchain经典
langchain-community==0.4.1 # langchain社区扩展
langchain-openai==1.1.10   # langchain的OpenAI集成
langchain-text-splitters==1.1.1 # 文本分割工具
langgraph==1.0.10       # LLM流程图
langgraph-checkpoint==4.0.1 # langgraph检查点
langgraph-prebuilt==1.0.8   # langgraph预构建
langgraph-sdk==0.3.9    # langgraph SDK
langsmith==0.7.10       # langchain调试工具
pinecone==7.3.0         # 向量数据库
pinecone-plugin-assistant==1.8.0 # pinecone助手插件
pinecone-plugin-interface==0.0.7 # pinecone接口插件
tiktoken==0.12.0        # OpenAI分词器
xxhash==3.6.0           # 高速哈希
zstandard==0.25.0       # 压缩算法

# ===========================
# 其他工具与扩展
# ===========================
propcache==0.4.1        # 属性缓存
uuid_utils==0.14.1      # UUID工具
python-dotenv==1.2.2    # 环境变量加载
click==8.3.1            # 命令行工具
jsonpatch==1.33         # JSON补丁
jsonpointer==3.0.0      # JSON指针

# ===========================
# 兼容性与辅助
# ===========================
annotated-types==0.7.0  # 类型注解扩展
annotated-doc==0.0.4    # 文档注解扩展